[#1398] fix(mr)(tez): Make attempId computable and move it to taskAttemptId in BlockId layout. (#1418)
### What changes were proposed in this pull request?

Before this PR, in MR and TEZ engine:
1. attemptId is in sequenceNo of BlockId instead of taskAttemptId.
2. attempId is fixed 6 bit.

After this PR:
1. attemptId is in taskAttemptId. This is more reasonable.
2. attempId is calculated from max num of allowed failures and whether speculative execution is enabled.

### Why are the changes needed?

Fix: #1398

### Does this PR introduce _any_ user-facing change?

No.

### How was this patch tested?
  
Existing UT and integrated tests.

+++ common/src/main/java/org/apache/uniffle/common/util/BlockId.java
@@ -32,10 +32,10 @@ public class BlockId {
   public final BlockIdLayout layout;
   public final int sequenceNo;
   public final int partitionId;
-  public final int taskAttemptId;
+  public final long taskAttemptId;
 
   protected BlockId(
-      long blockId, BlockIdLayout layout, int sequenceNo, int partitionId, int taskAttemptId) {
+      long blockId, BlockIdLayout layout, int sequenceNo, int partitionId, long taskAttemptId) {
     this.blockId = blockId;
     this.layout = layout;
     this.sequenceNo = sequenceNo;
